요청메소드 별 기능을 분리
GET, POST에 따른 동작과 관련된 메소드를 RequestHandler에서 GetMethodHandler, PostMethodHandler로 각각 이동시켰습니다.

RequestHandler의 run에서 선언된 httpResponse의 인스턴스를 매개로 전달받아 HTTP응답을 작성합니다.
